requirement.txt
websocket-client
python-telegram-bot==13.15
requests
pytz
pandas
python-docx